Create VLAD.FOSTER
Over a quarter of a million children in the United States are placed in the foster care system each year. Using a 2021 National Data Archive on Child Abuse and Neglect (NDACAN) with 105 variables and 631,893 observations, the child’s background, gender, race, age, length of stay, family structure, removal and discharge reasons were analyzed. In addition, whether the child was disabled can determine how long they stay in the foster care system. With the thought in mind of Every Child Deserves to Belong, the goals of this research is to minimize the length of stay of children in foster care and number of children in foster care.
Investigate why children from unmarried couples have the shortest length of stay in foster care
Investigating any possible improvement to the system for children with disabilities to transition to independence more quickly.
Investigating any possible improvement to the system for children with disabilities to transition to independence more quickly.
Investigating support for black females to reduce the number of children in foster care.
Educating all races about the opportunities to use the foster care system for education and training to transition to independence